+++ R/set_label.R
@@ -45,9 +45,11 @@
 #'
 #'
 #' # Set variable labels for data frame
-#' dummy <- data.frame(a = sample(1:4, 10, replace = TRUE),
-#'                     b = sample(1:4, 10, replace = TRUE),
-#'                     c = sample(1:4, 10, replace = TRUE))
+#' dummy <- data.frame(
+#'   a = sample(1:4, 10, replace = TRUE),
+#'   b = sample(1:4, 10, replace = TRUE),
+#'   c = sample(1:4, 10, replace = TRUE)
+#' )
 #' dummy <- set_label(dummy, c("Variable A", "Variable B", "Variable C"))
 #' str(dummy)
 #'
@@ -73,9 +75,11 @@
 #' # use 'var_labels()' to set labels within a pipe-workflow, and
 #' # when you need "tidyverse-consistent" api.
 #' # Set variable labels for data frame
-#' dummy <- data.frame(a = sample(1:4, 10, replace = TRUE),
-#'                     b = sample(1:4, 10, replace = TRUE),
-#'                     c = sample(1:4, 10, replace = TRUE))
+#' dummy <- data.frame(
+#'   a = sample(1:4, 10, replace = TRUE),
+#'   b = sample(1:4, 10, replace = TRUE),
+#'   c = sample(1:4, 10, replace = TRUE)
+#' )
 #'
 #' dummy %>%
 #'   var_labels(a = "First variable", c = "third variable") %>%