+++ R/zap_labels.R
@@ -111,7 +111,7 @@
 #' fill_labels(x)
 #' get_labels(fill_labels(x))
 #' # same as
-#' get_labels(x, include.non.labelled = TRUE)
+#' get_labels(x, non.labelled = TRUE)
 #'
 #' @importFrom stats na.omit
 #' @export